+++ src/card/cards/combat_constraints.rs
@@ -18,3 +18,4 @@ pub const ARMAGEDDON_CLOCK: CardDefinitionId = CardDefinitionId(1817);
 pub const OSAI_VULTURES: CardDefinitionId = CardDefinitionId(1818);
 pub const COCOON: CardDefinitionId = CardDefinitionId(1819);
 pub const VENARIAN_GOLD: CardDefinitionId = CardDefinitionId(1820);
+pub const ENERGY_TAP: CardDefinitionId = CardDefinitionId(1821);